+++ src/Stormancer.Plugins/GameFinder/Stormancer.Server.Plugins.GameFinder/Changelog.rst
@@ -16,6 +16,9 @@ Fixed
 - Fix code generator configuration to search GameFinder scene with the proper service type (gamefinder and not gamesession).
 - Add AreCompatible event handlers
 - Fix AreCompatible event handlers
+Added
+*****
+- Added LeaderPartyIsHost parameter to party gamefinder.
 
 7.0.0
 -----